I added the Document Provider branch for navigate from the old repository

+++ Owncloud iOs Client/DataBase/Queries/ManageFilesDB.h
@@ -37,6 +37,7 @@
  * Method that return an array of files, this files are sons of fileId
  *
  * @param fileId -> NSInteger
+
  *
  * @return list of files -> NSMutableArray
  *
@@ -278,7 +279,7 @@
  * @idFile -> id of the file to update the etag
  * @etag -> new etag
  */
-+ (void) updateEtagOfFileDtoByid:(int) idFile andNewEtag: (long long) etag;
++ (void) updateEtagOfFileDtoByid:(NSInteger) idFile andNewEtag: (long long) etag;
 
 /*
  * Method to update the etag of a file
